Implement loadFile plus changes required to implement

+++ internal/store/datafile.go
@@ -13,7 +13,7 @@ type DatafileSchema struct {
 	Description string            `db:"description" json:"description" r:"description"`
 	MediaType   DatafileMediaType `json:"mediatype" r:"mediatype"`
 	Parent      string            `db:"parent" json:"parent" r:"parent"`
-	Size        int               `db:"size" json:"size" r:"size"`
+	Size        int64             `db:"size" json:"size" r:"size"`
 	Uploaded    int               `db:"uploaded" json:"uploaded" r:"uploaded"`
 	UsesID      string            `db:"usesid" json:"usesid" r:"usesid"`
 }
@@ -30,7 +30,7 @@ type AddDatafileModel struct {
 	Description string `json:"description"`
 	UsesID      string
 	Parent      string
-	Size        int
+	Size        int64
 	ProjectID   string
 	DatadirID   string
 }
